safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.4_Expression_Statement/S12.4_A1-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Function statements must have a name.
S12.4_A1

PASS Expected parsing failure

TEST COMPLETE
